+++ packages/plugin-packages/mashroom-http-proxy/test/proxy/ProxyImplNodeHttpProxy.pure.test.ts
@@ -436,7 +436,9 @@ describe('ProxyImplNodeHttpProxy', () => {
         req.headers.upgrade = 'whatever';
         const head = Buffer.from('');
 
-        const pluginRegistry: any = {};
+        const pluginRegistry: any = {
+            interceptors: [],
+        };
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
         const headerFilter = new HttpHeaderFilter([]);
         const httpProxy = new ProxyImplNodeHttpProxy(2000, false, interceptorHandler, headerFilter, false, null, null, null, false, loggingUtils.dummyLoggerFactory);